utils: export WaitObserver methods into wait package
