+++ internal/utils/utils.go
@@ -4,12 +4,10 @@
 package utils
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"regexp"
 	"strings"
-	"time"
 
 	"github.com/blang/semver/v4"
 
@@ -91,78 +89,6 @@ func BuildImagePath(userImage, userVersion, defaultImage, defaultVersion string,
 		}
 	}
 	return image
-}
-
-type LogFunc func(err error, waitTime string)
-
-type WaitParameters struct {
-	RetryInterval   time.Duration
-	WarningInterval time.Duration
-	Timeout         time.Duration
-	Log             LogFunc
-}
-
-func (w WaitParameters) retryInterval() time.Duration {
-	if w.RetryInterval != time.Duration(0) {
-		return w.RetryInterval
-	}
-
-	return defaults.WaitRetryInterval
-}
-
-func (w WaitParameters) warningInterval() time.Duration {
-	if w.WarningInterval != time.Duration(0) {
-		return w.WarningInterval
-	}
-
-	return defaults.WaitWarningInterval
-}
-
-type WaitObserver struct {
-	ctx         context.Context
-	params      WaitParameters
-	lastWarning time.Time
-	waitStarted time.Time
-	cancel      context.CancelFunc
-}
-
-func NewWaitObserver(ctx context.Context, p WaitParameters) *WaitObserver {
-	w := &WaitObserver{
-		ctx:         ctx,
-		params:      p,
-		waitStarted: time.Now(),
-	}
-
-	if p.Timeout != time.Duration(0) {
-		w.ctx, w.cancel = context.WithTimeout(ctx, p.Timeout)
-	}
-
-	return w
-}
-
-func (w *WaitObserver) Cancel() {
-	if w.cancel != nil {
-		w.cancel()
-	}
-}
-
-func (w *WaitObserver) Retry(err error) error {
-	if w.params.Log != nil && time.Since(w.lastWarning) > w.params.warningInterval() {
-		waitString := time.Since(w.waitStarted).Truncate(time.Second).String()
-		w.params.Log(err, waitString)
-		w.lastWarning = time.Now()
-	}
-
-	select {
-	case <-w.ctx.Done():
-		if err != nil {
-			return fmt.Errorf("timeout while waiting for condition, last error: %s", err)
-		}
-		return fmt.Errorf("timeout while waiting for condition")
-	case <-time.After(w.params.retryInterval()):
-	}
-
-	return nil
 }
 
 func Contains(l []string, v string) bool {
